rename disposable to destroyable
