+++ libs/client/common/src/lib/extends/index.ts
@@ -1,1 +1,1 @@
-export * from './disposable'
+export * from './destroyable'
